accounting_system_windows/requirements.txt
PyQt6>=6.4.0
SQLAlchemy>=2.0.0
PyYAML>=6.0
Pydantic>=2.0.0
cryptography>=41.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
reportlab>=4.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
